Mejorada la seguridad de la API para imprimir tickets.
git-svn-id: http://facturascripts.googlecode.com/svn/branches/0.9@798 fe615fe8-962f-0410-811c-9ded47c8d053

+++ api.php
@@ -32,16 +32,34 @@ else
    $db = new fs_postgresql();
 }
 
+require_once 'base/fs_model.php';
+require_model('caja.php');
+
 if( $db->connect() )
 {
    if( isset($_GET['remote-printer']) )
    {
       if(FS_PRINTER == 'remote-printer')
       {
-         if( file_exists('tmp/'.FS_TMP_NAME.'remote-printer.txt') )
+         /**
+          * Añadimos un poquito de seguridad.
+          * Comprobamos que la IP desde la que se quiere imprimir corresponda
+          * con la del usuario que ha abierto la caja.
+          */
+         $caja = new caja();
+         $caja0 = $caja->get_last_from_this_server();
+         if( $caja0 AND isset($_SERVER['REMOTE_ADDR']) )
          {
-            echo file_get_contents('tmp/'.FS_TMP_NAME.'remote-printer.txt');
-            unlink('tmp/'.FS_TMP_NAME.'remote-printer.txt');
+            if( $caja0->ip == $_SERVER['REMOTE_ADDR'] OR is_null($caja0->ip) )
+            {
+               if( file_exists('tmp/'.FS_TMP_NAME.'remote-printer.txt') )
+               {
+                  echo file_get_contents('tmp/'.FS_TMP_NAME.'remote-printer.txt');
+                  unlink('tmp/'.FS_TMP_NAME.'remote-printer.txt');
+               }
+            }
+            else
+               echo 'ERROR';
          }
       }
       else
